RF-23096 Properly set --custom-id flag when parameter is set
Unlike in the GitHub Action, we don't do anything fancy with the --environmnent-id flag: if you set both parameters, both flags will be passed to the CLI (and the CLI will ignore the `--environment-id` flag). We log a warning message in the Orb in that case to help you debug.